requirements.txt
numpy # BSD License
python-dotenv # BSD License (BSD-3-Clause)
matplotlib # Python Software Foundation License
japanize-matplotlib
streamlit==1.34.0 # Apache-2.0 license
fastapi==0.109.2 # MIT license
gunicorn==21.2.0 # MIT license
uvicorn[standard]==0.27.1 # BSD License

rasterio==1.4.2 # BSD License
geopandas==1.0.1 # BSD-3-Clause license
shapely==2.0.6 # BSD-3-Clause license
overpy==0.7 # MIT license

google-auth==2.27.0
google-cloud-aiplatform==1.71.1

pandas
requests

